+++ test/large_number_test.dart
@@ -1,58 +1,7 @@
 import 'dart:ffi';
-import 'package:ffi/ffi.dart';
 import 'package:test/test.dart';
 import 'package:wincalc_engine/wincalc_engine.dart';
-
-/// Helper function to get string result from native buffer
-String getResultInRadix(Pointer<CalculatorInstance> instance, int radix) {
-  const bufferSize = 256;
-  final buffer = calloc<Char>(bufferSize);
-  try {
-    switch (radix) {
-      case 16:
-        calculator_get_result_hex(instance, buffer, bufferSize);
-        break;
-      case 10:
-        calculator_get_result_dec(instance, buffer, bufferSize);
-        break;
-      case 8:
-        calculator_get_result_oct(instance, buffer, bufferSize);
-        break;
-      case 2:
-        calculator_get_result_bin(instance, buffer, bufferSize);
-        break;
-      default:
-        throw ArgumentError('Invalid radix: $radix');
-    }
-    return buffer.cast<Utf8>().toDartString();
-  } finally {
-    calloc.free(buffer);
-  }
-}
-
-/// Helper function to send a hex digit (0-F)
-void sendHexDigit(Pointer<CalculatorInstance> instance, int digit) {
-  final command = switch (digit) {
-    0 => CMD_0,
-    1 => CMD_1,
-    2 => CMD_2,
-    3 => CMD_3,
-    4 => CMD_4,
-    5 => CMD_5,
-    6 => CMD_6,
-    7 => CMD_7,
-    8 => CMD_8,
-    9 => CMD_9,
-    10 => CMD_A,
-    11 => CMD_B,
-    12 => CMD_C,
-    13 => CMD_D,
-    14 => CMD_E,
-    15 => CMD_F,
-    _ => throw ArgumentError('Invalid hex digit: $digit'),
-  };
-  calculator_send_command(instance, command);
-}
+import 'test_helpers.dart';
 
 void main() {
   late Pointer<CalculatorInstance> calc;